refactor: Export Viewer helpers in helpers file
