+++ react/Viewer/index.jsx
@@ -1,50 +1,18 @@
 import React, { Component, createRef } from 'react'
 import PropTypes from 'prop-types'
 
-import { isMobile as isMobileDevice } from 'cozy-device-helper'
-
 import withBreakpoints from '../helpers/withBreakpoints'
 import { FileDoctype } from '../proptypes'
 
 import ViewerWrapper from './ViewerWrapper'
 import ViewerControls from './ViewerControls'
-import ImageViewer from './ImageViewer'
-import AudioViewer from './AudioViewer'
-import VideoViewer from './VideoViewer'
-import PdfJsViewer from './PdfJsViewer'
-import TextViewer from './TextViewer'
-import PdfMobileViewer from './PdfMobileViewer'
-import NoViewer from './NoViewer'
-import ShortcutViewer from './ShortcutViewer'
 import InformationPanel from './InformationPanel'
 import Footer from './Footer'
+import { getViewerComponentName } from './helpers'
 
 const KEY_CODE_LEFT = 37
 const KEY_CODE_RIGHT = 39
 const KEY_CODE_ESCAPE = 27
-
-export const isPlainText = (mimeType = '', fileName = '') => {
-  return mimeType ? /^text\//.test(mimeType) : /\.(txt|md)$/.test(fileName)
-}
-
-export const getViewerComponentName = (file, isDesktop) => {
-  switch (file.class) {
-    case 'shortcut':
-      return ShortcutViewer
-    case 'image':
-      return ImageViewer
-    case 'audio':
-      return AudioViewer
-    case 'video':
-      return isMobileDevice() ? NoViewer : VideoViewer
-    case 'pdf':
-      return isDesktop ? PdfJsViewer : PdfMobileViewer
-    case 'text':
-      return isPlainText(file.mime, file.name) ? TextViewer : NoViewer
-    default:
-      return NoViewer
-  }
-}
 
 export class Viewer extends Component {
   constructor() {
